Move sgs_policies and gphh to do_solver
- sgs_policies is not a solver itself and is used only by gphh and
do_solver
- do_solver and gphh are both solvers based on discrete-optimization
- gphh is to be using discrete_optimization's own GPHH

=> we put stuff related to d-o into do_solver subpackage
