+++ examples/scheduling/gphh_example.py
@@ -10,7 +10,7 @@ from skdecide import rollout_episode
 from skdecide.hub.domain.rcpsp.rcpsp_sk import RCPSP
 from skdecide.hub.domain.rcpsp.rcpsp_sk_parser import load_domain
 from skdecide.hub.solver.do_solver.do_solver_scheduling import DOSolver, SolvingMethod
-from skdecide.hub.solver.gphh.gphh import (
+from skdecide.hub.solver.do_solver.gphh import (
     GPHH,
     EvaluationGPHH,
     FeatureEnum,
@@ -29,7 +29,7 @@ from skdecide.hub.solver.gphh.gphh import (
     min_operator,
     protected_div,
 )
-from skdecide.hub.solver.sgs_policies.sgs_policies import (
+from skdecide.hub.solver.do_solver.sgs_policies import (
     BasePolicyMethod,
     PolicyMethodParams,
 )
