+++ skdecide/hub/solver/do_solver/__init__.py
@@ -3,3 +3,15 @@
 # LICENSE file in the root directory of this source tree.
 
 from .do_solver_scheduling import DOSolver
+from .gphh import (
+    GPHH,
+    EvaluationGPHH,
+    FeatureEnum,
+    FixedPermutationPolicy,
+    GPHHPolicy,
+    ParametersGPHH,
+    PermutationDistance,
+    PoolAggregationMethod,
+    PooledGPHHPolicy,
+)
+from .sgs_policies import BasePolicyMethod, PolicyMethodParams, PolicyRCPSP
