+++ skdecide/hub/solver/gphh/__init__.py
@@ -1,15 +1,3 @@
 # Copyright (c) AIRBUS and its affiliates.
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
-
-from .gphh import (
-    GPHH,
-    EvaluationGPHH,
-    FeatureEnum,
-    FixedPermutationPolicy,
-    GPHHPolicy,
-    ParametersGPHH,
-    PermutationDistance,
-    PoolAggregationMethod,
-    PooledGPHHPolicy,
-)
