+++ skdecide/hub/solver/sgs_policies/__init__.py
@@ -1,5 +1,3 @@
 # Copyright (c) AIRBUS and its affiliates.
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
-
-from .sgs_policies import BasePolicyMethod, PolicyMethodParams, PolicyRCPSP
